Factor out to_string, making Base hold what generic uses.
