+++ astropy/units/format/console.py
@@ -33,23 +33,11 @@ class Console(base.Base):
         return m
 
     @classmethod
+    def _format_superscript(cls, number):
+        return f"^{number}"
+
+    @classmethod
     def format_exponential_notation(cls, val, format_spec=".8g"):
-        """
-        Formats a value in exponential notation.
-
-        Parameters
-        ----------
-        val : number
-            The value to be formatted
-
-        format_spec : str, optional
-            Format used to split up mantissa and exponent
-
-        Returns
-        -------
-        string : str
-            The value in exponential notation in a this class's format.
-        """
         m, ex = utils.split_mantissa_exponent(val, format_spec)
 
         parts = []
@@ -76,34 +64,5 @@ class Console(base.Base):
 
     @classmethod
     def to_string(cls, unit, inline=True):
-        if unit.scale == 1:
-            s = ""
-        else:
-            # Non-unity scale happens mostly for decomposed units.
-            # E.g., u.Ry.decompose() gives "2.17987e-18 kg m2 / s2".
-            s = cls.format_exponential_notation(unit.scale)
-
-        # Take care that dimensionless does not have bases (but can
-        # have a scale; e.g., u.percent.decompose() gives "0.01").
-        if len(unit.bases):
-            if s:
-                s += cls._space
-            if inline:
-                nominator = zip(unit.bases, unit.powers)
-                denominator = []
-            else:
-                nominator, denominator = utils.get_grouped_by_powers(
-                    unit.bases, unit.powers
-                )
-            if len(denominator):
-                if len(nominator):
-                    nominator = cls._format_unit_list(nominator)
-                else:
-                    nominator = "1"
-                denominator = cls._format_unit_list(denominator)
-                s = cls._format_fraction(s, nominator, denominator)
-            else:
-                nominator = cls._format_unit_list(nominator)
-                s += nominator
-
-        return s
+        # Change default of inline to True.
+        return super().to_string(unit, inline=inline)
